+++ src/cart/index.html
@@ -10,9 +10,7 @@
   </head>
 
   <body>
-    <header class="divider" id="header-main">
-      
-    </header>
+    <header class="divider" id="main-header"></header>
 
     <main class="divider">
       <section class="products">
@@ -56,6 +54,6 @@
       </div>
     </main>
 
-    <footer id="footer-main"></footer>
+    <footer id="main-footer"></footer>
   </body>
 </html>
