+++ src/checkout/index.html
@@ -9,9 +9,7 @@
   </head>
 
   <body>
-    <header class="divider" id="header-main">
-      
-    </header>
+    <header class="divider" id="header-main"></header>
 
     <main class="divider">
       <section class="products">
