+++ src/js/main.js
@@ -1,13 +1,10 @@
+import { initCartIcon } from "./Cart.mjs";
 import ProductData from "./ProductData.mjs";
 import ProductListing from "./ProductList.mjs";
-import { initCartIcon } from "./Cart.mjs";
 import { loadHeaderFooter, qs } from "./utils.mjs";
 
 //call header and footer from the template
-loadHeaderFooter();
-
-// show number of items in cart on the cart icon in header
-initCartIcon();
+loadHeaderFooter(null, initCartIcon);
 
 // show the list of products
 var products = new ProductData("tents");
